@@ -1,8 +1,9 @@
 {
   "name": "OctoLinker",
-  "version": "6.8.1",
+  "version": "0.0.0",
   "engines": {
-    "node": ">=12.0.0"
+    "node": ">= 16",
+    "npm": ">= 7"
   },
   "scripts": {
     "lint": "eslint .",
@@ -10,9 +11,8 @@
     "test": "jest",
     "e2e": "node ./e2e/generate-fixtures-json.js && jest --config=e2e/jest-e2e.config.js",
     "test:watch": "jest --watch",
-    "version": "./scripts/version.sh",
-    "release:cws": "cd dist && webstore upload --auto-publish",
-    "release:amo": "cd dist && web-ext-submit",
+    "release:cws": "cd dist && npx chrome-webstore-upload-cli@next upload --auto-publish",
+    "release:amo": "cd dist && npx web-ext-submit@5",
     "release": "yarn release:cws && yarn release:amo",
     "build": "webpack --mode=production",
     "start": "webpack --watch --mode=development",
@@ -25,49 +25,45 @@
     "safari-launch": "open 'safari/build/Release/OctoLinker.app'"
   },
   "devDependencies": {
-    "@babel/core": "^7.14.6",
-    "@babel/plugin-proposal-object-rest-spread": "^7.14.5",
-    "@babel/plugin-transform-react-jsx": "^7.14.5",
-    "@babel/plugin-transform-runtime": "^7.14.5",
-    "@babel/preset-env": "^7.14.7",
+    "@babel/core": "^7.18.10",
+    "@babel/eslint-parser": "^7.18.9",
+    "@babel/plugin-transform-react-jsx": "^7.18.10",
+    "@babel/plugin-transform-runtime": "^7.18.10",
+    "@babel/preset-env": "^7.18.10",
     "assert": "^2.0.0",
-    "async": "^3.2.0",
+    "async": "^3.2.4",
     "babel-core": "^7.0.0-bridge.0",
-    "babel-eslint": "^10.1.0",
-    "babel-jest": "^27.0.6",
-    "babel-loader": "^8.2.2",
-    "babel-preset-jest": "^27.0.1",
+    "babel-jest": "^28.1.1",
+    "babel-loader": "^8.2.5",
+    "babel-preset-jest": "^28.0.2",
     "chrome-launch": "^1.1.4",
-    "chrome-webstore-upload-cli": "^1.2.1",
-    "copy-webpack-plugin": "^9.0.1",
-    "css-loader": "^5.2.6",
+    "copy-webpack-plugin": "^10.2.4",
+    "css-loader": "^6.7.1",
     "dot-json": "^1.2.2",
-    "dotenv": "^10.0.0",
-    "eslint": "^7.30.0",
-    "eslint-config-airbnb": "^18.2.1",
-    "eslint-config-prettier": "^8.3.0",
-    "eslint-plugin-import": "^2.23.4",
-    "eslint-plugin-jsx-a11y": "^6.4.1",
-    "eslint-plugin-prettier": "^3.4.0",
-    "eslint-plugin-react": "^7.24.0",
-    "jest": "^26.6.3",
+    "dotenv": "^16.0.1",
+    "eslint": "^8.22.0",
+    "eslint-config-airbnb": "^19.0.4",
+    "eslint-config-prettier": "^8.5.0",
+    "eslint-plugin-import": "^2.26.0",
+    "eslint-plugin-jsx-a11y": "^6.5.1",
+    "eslint-plugin-prettier": "^4.2.1",
+    "eslint-plugin-react": "^7.30.1",
+    "jest": "^27.5.1",
     "jest-fetch-mock": "^3.0.3",
-    "jest-puppeteer": "^5.0.4",
+    "jest-puppeteer": "^6.1.1",
     "json-loader": "^0.5.7",
     "path-browserify": "^1.0.1",
-    "prettier": "^2.2.1",
+    "prettier": "^2.7.1",
     "process": "^0.11.10",
-    "puppeteer": "^9.1.1",
+    "puppeteer": "^16.2.0",
     "recursive-readdir": "^2.2.2",
-    "sinon": "^9.2.4",
     "stream-browserify": "^3.0.0",
-    "style-loader": "^2.0.0",
+    "style-loader": "^3.3.1",
     "url": "^0.11.0",
     "vm-browserify": "^1.1.2",
-    "web-ext": "^6.2.0",
-    "web-ext-submit": "^6.2.0",
-    "webpack": "^5.42.0",
-    "webpack-cli": "^4.7.2",
+    "web-ext": "^6.8.0",
+    "webpack": "^5.74.0",
+    "webpack-cli": "^4.10.0",
     "webpack-extension-reloader": "^1.1.4"
   },
   "optionalDependencies": {
