@@ -1,10 +1,10 @@
-require('dotenv').config();
+import 'dotenv/config';
 
-const fixtures = require('./fixtures.json'); // eslint-disable-line import/no-unresolved
-const diffFixtures = require('./diff-fixtures.json'); // eslint-disable-line import/no-unresolved
+import fixtures from './fixtures.json'; // eslint-disable-line import/no-unresolved
+import diffFixtures from './diff-fixtures.json';
 
 async function executeTest(url, targetUrl, selector) {
-  if ((await page.url()) !== url) {
+  if (page.url() !== url) {
     await page.goto(url);
   }
 
@@ -45,7 +45,7 @@ describe('End to End tests', () => {
       throw new Error(authError);
     } catch (error) {
       if (!error.message.includes('failed to find element matching selector')) {
-        await expect(error).toBeUndefined();
+        expect(error).toBeUndefined();
       }
     }
 
@@ -93,25 +93,25 @@ describe('End to End tests', () => {
     });
   });
 
-  describe('diff view', () => {
-    diffFixtures.forEach(({ url, targetUrl }) => {
-      it(`resolves ${url} to ${targetUrl}`, async () => {
-        await executeTest(
-          url,
-          targetUrl,
-          '.selected-line.blob-code .octolinker-link',
-        );
-      });
-    });
+  // describe('diff view', () => {
+  //   diffFixtures.forEach(({ url, targetUrl }) => {
+  //     it(`resolves ${url} to ${targetUrl}`, async () => {
+  //       await executeTest(
+  //         url,
+  //         targetUrl,
+  //         '.selected-line.blob-code .octolinker-link',
+  //       );
+  //     });
+  //   });
 
-    it('links commentbox', async () => {
-      await executeTest(
-        'https://github.com/OctoLinker/OctoLinker/pull/451/files',
-        'https://nodejs.org/api/path.html',
-        '.highlight-source-js .octolinker-link',
-      );
-    });
-  });
+  //   it('links commentbox', async () => {
+  //     await executeTest(
+  //       'https://github.com/OctoLinker/OctoLinker/pull/451/files',
+  //       'https://nodejs.org/api/path.html',
+  //       '.highlight-source-js .octolinker-link',
+  //     );
+  //   });
+  // });
 
   describe.skip('expanded blob', () => {
     it('should resolve after appending new blobs', async () => {
