@@ -4,5 +4,5 @@ import java.util.List;
 // @OctoLinkerResolve(http://fasterxml.github.io/jackson-databind/javadoc/2.9/com/fasterxml/jackson/databind/JsonNode.html)
 import com.fasterxml.jackson.databind.JsonNode;
 
-// @OctoLinkerResolve(https://docs.spring.io/spring/docs/5.1.7.RELEASE/javadoc-api/org/springframework/stereotype/Component.html)
+// @OctoLinkerResolve(https://docs.spring.io/spring-framework/docs/current/javadoc-api/org/springframework/stereotype/Component.html)
 import org.springframework.stereotype.Component;
