@@ -1,6 +1,9 @@
 // @OctoLinkerResolve(https://nodejs.org/api/fs.html)
 require('fs');
 
+// @OctoLinkerResolve(https://nodejs.org/api/fs.html)
+require('node:fs');
+
 // @OctoLinkerResolve(https://github.com/lodash/lodash)
 require("lodash");
 
@@ -10,11 +13,19 @@ require('react');
 // @OctoLinkerResolve(https://nodejs.org/api/http.html)
 const http = require('http');
 
+// @OctoLinkerResolve(https://nodejs.org/api/http.html)
+const http2 = require('node:http');
+
 // @OctoLinkerResolve(https://nodejs.org/api/cluster.html)
 const { cluster } = require('cluster');
 
+const { cluster: cluster2 } = require('node:cluster');
+
 // @OctoLinkerResolve(https://nodejs.org/api/path.html)
 import path from "path";
+
+// @OctoLinkerResolve(https://nodejs.org/api/fs.html)
+import fs from "node:fs";
 
 // @OctoLinkerResolve(https://nodejs.org/api/url.html)
 import url from 'url';
@@ -25,9 +36,17 @@ import * as fs from "https://deno.land/x/std/fs/mod.ts";
 // @OctoLinkerResolve(https://nodejs.org/api/os.html)
 import { os } from 'os';
 
-import { 
-  foo 
+// @OctoLinkerResolve(https://nodejs.org/api/os.html)
+import { os as os2 } from 'node:os';
+
+import {
+  foo
 } from 'dns';
+// @OctoLinkerResolveAbove(https://nodejs.org/api/dns.html)
+
+import {
+  foo2
+} from 'node:dns';
 // @OctoLinkerResolveAbove(https://nodejs.org/api/dns.html)
 
 // @OctoLinkerResolve(https://nodejs.org/api/net.html)
