@@ -59,7 +59,7 @@ describe('helper-github-api tree', () => {
   });
 
   it('returns an empty array on failure', async () => {
-    fetch.mockResponseOnce(() => new Promise((resolve, reject) => reject()));
+    fetch.mockResponseOnce(() => Promise.reject());
 
     expect(await fetchTree(options)).toEqual([]);
   });
