@@ -470,6 +470,12 @@ const fixtures = {
       ['import org.hamcrest.Foo', ['org.hamcrest.Foo']],
       ['import com.fasterxml.Foo', ['com.fasterxml.Foo']],
       ['import org.junit.Foo', ['org.junit.Foo']],
+      ['import lombok.Foo', ['lombok.Foo']],
+      ['import org.atteo.Foo', ['org.atteo.Foo']],
+      ['import org.objectweb.Foo', ['org.objectweb.Foo']],
+      ['import org.hamcrest.Foo', ['org.hamcrest.Foo']],
+      ['import ch.qos.logback.Foo', ['ch.qos.logback.Foo']],
+      ['import yapion.Foo', ['yapion.Foo']],
       ['import org.mockito.Foo', ['org.mockito.Foo']],
       ['import org.apache.Foo', ['org.apache.Foo']],
       ['import org.slf4j.Foo', ['org.slf4j.Foo']],
@@ -522,8 +528,12 @@ const fixtures = {
   GITHUB_ACTIONS: {
     valid: [
       ['uses: foo/bar', ['foo/bar']],
-      ['uses: foo/bar@v1', ['foo/bar']],
-      ['uses: foo/bar@master', ['foo/bar']],
+      ['uses: foo/bar@v1', ['foo/bar', 'v1']],
+      ['uses: foo/bar@master', ['foo/bar', 'master']],
+      ['uses: "foo/bar"', ['foo/bar']],
+      ['uses: "foo/bar@v1"', ['foo/bar', 'v1']],
+      ["uses: 'foo/bar'", ['foo/bar']],
+      ["uses: 'foo/bar@v1'", ['foo/bar', 'v1']],
     ],
     invalid: [],
   },
