@@ -1,6 +1,6 @@
 import { h } from 'preact';
 
-export default ({
+export function Input({
   name,
   label,
   description,
@@ -8,22 +8,24 @@ export default ({
   onInput,
   className,
   type = 'text',
-}) => (
-  <p>
-    <label htmlFor={name} id={name}>
-      <strong>{label}</strong>
-    </label>
+}) {
+  return (
+    <p>
+      <label htmlFor={name} id={name}>
+        <strong>{label}</strong>
+      </label>
 
-    <input
-      style={{ width: '100%' }}
-      className={className}
-      type={type}
-      id={name}
-      name={name}
-      value={value}
-      onInput={onInput}
-    />
+      <input
+        style={{ width: '100%' }}
+        className={className}
+        type={type}
+        id={name}
+        name={name}
+        value={value}
+        onInput={onInput}
+      />
 
-    <div className="note">{description}</div>
-  </p>
-);
+      <div className="note">{description}</div>
+    </p>
+  );
+}
